Fix: Move exception to Extension namespace
